+++ toolium/driver_wrappers_pool.py
@@ -166,21 +166,21 @@ class DriverWrappersPool(object):
                 os.makedirs(cls.output_directory)
 
     @classmethod
-    def configure_visual_directories(cls, browser_info):
+    def configure_visual_directories(cls, driver_info):
         """Configure screenshots, videos and visual directories
 
-        :param browser_info: browser property value to rename folders
+        :param driver_info: driver property value to rename folders
         """
         if cls.screenshots_directory is None:
             # Unique screenshots and videos directories
             date = datetime.datetime.now().strftime('%Y-%m-%d_%H%M%S')
-            cls.screenshots_directory = os.path.join(cls.output_directory, 'screenshots', date + '_' + browser_info)
+            cls.screenshots_directory = os.path.join(cls.output_directory, 'screenshots', date + '_' + driver_info)
             cls.screenshots_number = 1
-            cls.videos_directory = os.path.join(cls.output_directory, 'videos', date + '_' + browser_info)
+            cls.videos_directory = os.path.join(cls.output_directory, 'videos', date + '_' + driver_info)
             cls.videos_number = 1
 
             # Unique visualtests directories
-            cls.visual_output_directory = os.path.join(cls.output_directory, 'visualtests', date + '_' + browser_info)
+            cls.visual_output_directory = os.path.join(cls.output_directory, 'visualtests', date + '_' + driver_info)
             cls.visual_number = 1
 
     @classmethod