+++ toolium/test/test_driver_wrapper.py
@@ -39,7 +39,7 @@ class DriverWrapperCommon(unittest.TestCase):
         try:
             del os.environ["Config_log_filename"]
             del os.environ["Config_prop_filenames"]
-            del os.environ["Browser_browser"]
+            del os.environ["Driver_type"]
         except KeyError:
             pass
 
@@ -64,14 +64,14 @@ class DriverWrapperPropertiesTests(DriverWrapperCommon):
     def test_configure_properties(self):
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         self.driver_wrapper.configure_properties()
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))  # get last value
         assert_equal('5', self.driver_wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
         assert_equal(None, self.driver_wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_android(self):
         os.environ["Config_prop_filenames"] = 'android-properties.cfg'
         self.driver_wrapper.configure_properties()
-        assert_equal('android', self.driver_wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('android', self.driver_wrapper.config.get('Driver', 'type'))  # get last value
         assert_equal(None, self.driver_wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
         assert_equal('http://invented_url/Demo.apk',
                      self.driver_wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
@@ -79,7 +79,7 @@ class DriverWrapperPropertiesTests(DriverWrapperCommon):
     def test_configure_properties_two_files(self):
         os.environ["Config_prop_filenames"] = 'properties.cfg;android-properties.cfg'
         self.driver_wrapper.configure_properties()
-        assert_equal('android', self.driver_wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('android', self.driver_wrapper.config.get('Driver', 'type'))  # get last value
         assert_equal('5', self.driver_wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
         assert_equal('http://invented_url/Demo.apk',
                      self.driver_wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
@@ -87,20 +87,20 @@ class DriverWrapperPropertiesTests(DriverWrapperCommon):
     def test_configure_properties_two_files_android_first(self):
         os.environ["Config_prop_filenames"] = 'android-properties.cfg;properties.cfg'
         self.driver_wrapper.configure_properties()
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))  # get last value
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))  # get last value
         assert_equal('5', self.driver_wrapper.config.get_optional('Common', 'implicitly_wait'))  # only in properties
         assert_equal('http://invented_url/Demo.apk',
                      self.driver_wrapper.config.get_optional('AppiumCapabilities', 'app'))  # only in android
 
     def test_configure_properties_system_property(self):
         os.environ["Config_prop_filenames"] = 'properties.cfg'
-        os.environ["Browser_browser"] = 'opera'
+        os.environ["Driver_type"] = 'opera'
         self.driver_wrapper.configure_properties()
-        assert_equal('opera', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('opera', self.driver_wrapper.config.get('Driver', 'type'))
 
     def test_configure_properties_file_default_file(self):
         self.driver_wrapper.configure_properties()
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))
 
     def test_configure_properties_file_not_found(self):
         os.environ["Config_prop_filenames"] = 'notfound-properties.cfg'
@@ -112,36 +112,36 @@ class DriverWrapperPropertiesTests(DriverWrapperCommon):
         # Configure properties
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         self.driver_wrapper.configure_properties()
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))
 
         # Modify property
-        new_browser = 'opera'
-        self.driver_wrapper.config.set('Browser', 'browser', new_browser)
-        assert_equal(new_browser, self.driver_wrapper.config.get('Browser', 'browser'))
+        new_driver_type = 'opera'
+        self.driver_wrapper.config.set('Driver', 'type', new_driver_type)
+        assert_equal(new_driver_type, self.driver_wrapper.config.get('Driver', 'type'))
 
         # Trying to configure again
         self.driver_wrapper.configure_properties()
 
         # Configuration has not been initialized
-        assert_equal(new_browser, self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal(new_driver_type, self.driver_wrapper.config.get('Driver', 'type'))
 
     def test_configure_properties_change_configuration_file(self):
         # Configure properties
         os.environ["Config_prop_filenames"] = 'properties.cfg'
         self.driver_wrapper.configure_properties()
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))
 
         # Modify property
-        new_browser = 'opera'
-        self.driver_wrapper.config.set('Browser', 'browser', new_browser)
-        assert_equal(new_browser, self.driver_wrapper.config.get('Browser', 'browser'))
+        new_driver_type = 'opera'
+        self.driver_wrapper.config.set('Driver', 'type', new_driver_type)
+        assert_equal(new_driver_type, self.driver_wrapper.config.get('Driver', 'type'))
 
         # Change file and try to configure again
         os.environ["Config_prop_filenames"] = 'android-properties.cfg'
         self.driver_wrapper.configure_properties()
 
         # Check that configuration has been initialized
-        assert_equal('android', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('android', self.driver_wrapper.config.get('Driver', 'type'))
 
 
 class DriverWrapperLoggerTests(DriverWrapperCommon):
@@ -227,8 +227,8 @@ class DriverWrapperBaselineTests(DriverWrapperCommon):
         assert_equal(self.driver_wrapper.baseline_name, 'firefox')
         assert_equal(self.driver_wrapper.visual_baseline_directory, expected_baseline_directory)
 
-    def test_configure_visual_baseline_browser(self):
-        self.driver_wrapper.config.set('VisualTests', 'baseline_name', '{Browser_browser}')
+    def test_configure_visual_baseline_driver_type(self):
+        self.driver_wrapper.config.set('VisualTests', 'baseline_name', '{Driver_type}')
 
         self.driver_wrapper.configure_visual_baseline()
 
@@ -274,7 +274,7 @@ class DriverWrapperBaselineTests(DriverWrapperCommon):
         assert_equal(self.driver_wrapper.visual_baseline_directory, expected_baseline_directory)
 
 
-# (browser, is_mobile, is_android, is_ios)
+# (driver_type, is_mobile, is_android, is_ios)
 mobile_tests = (
     ('android-4.1.2-on-android', True, True, False),
     ('android', True, True, False),
@@ -284,7 +284,7 @@ mobile_tests = (
     ('firefox', False, False, False),
 )
 
-# (browser, appium_app, appium_browser_name, is_web, is_android_web, is_ios_web)
+# (driver_type, appium_app, appium_browser_name, is_web, is_android_web, is_ios_web)
 web_tests = (
     ('android-4.1.2-on-android', 'C:/Demo.apk', None, False, False, False),
     ('android', 'C:/Demo.apk', None, False, False, False),
@@ -302,8 +302,8 @@ web_tests = (
     ('firefox', None, None, True, False, False),
 )
 
-# (browser, is_maximizable)
-maximizable_browsers = (
+# (driver_type, is_maximizable)
+maximizable_drivers = (
     ('firefox-4.1.2-on-android', True),
     ('firefox', True),
     ('opera-12.12-on-xp', True),
@@ -328,41 +328,41 @@ class DriverWrapperTests(DriverWrapperCommon):
         new_wrapper = DriverWrapper()
 
         # Modify new wrapper
-        first_browser = 'firefox'
-        new_browser = 'opera'
-        new_wrapper.config.set('Browser', 'browser', new_browser)
+        first_driver_type = 'firefox'
+        new_driver_type = 'opera'
+        new_wrapper.config.set('Driver', 'type', new_driver_type)
 
         # Check that wrapper and new_wrapper are different
-        assert_equal(first_browser, self.driver_wrapper.config.get('Browser', 'browser'))
-        assert_equal(new_browser, new_wrapper.config.get('Browser', 'browser'))
+        assert_equal(first_driver_type, self.driver_wrapper.config.get('Driver', 'type'))
+        assert_equal(new_driver_type, new_wrapper.config.get('Driver', 'type'))
         assert_not_equal(self.driver_wrapper, new_wrapper)
 
     def test_configure_no_changes(self):
         # Check previous values
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))
 
         # Modify wrapper
-        self.driver_wrapper.config.set('Browser', 'browser', 'opera')
+        self.driver_wrapper.config.set('Driver', 'type', 'opera')
 
         # Trying to configure again
         self.driver_wrapper.configure()
 
         # Configuration has not been initialized
-        assert_equal('opera', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('opera', self.driver_wrapper.config.get('Driver', 'type'))
 
     def test_configure_change_configuration_file(self):
         # Check previous values
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))
 
         # Modify wrapper
-        self.driver_wrapper.config.set('Browser', 'browser', 'opera')
+        self.driver_wrapper.config.set('Driver', 'type', 'opera')
 
-        # Change browser and try to configure again
+        # Change driver type and try to configure again
         os.environ["Config_prop_filenames"] = os.path.join(self.root_path, 'conf', 'android-properties.cfg')
         self.driver_wrapper.configure()
 
         # Check that configuration has been initialized
-        assert_equal('android', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('android', self.driver_wrapper.config.get('Driver', 'type'))
 
     @mock.patch('toolium.driver_wrapper.ConfigDriver')
     def test_connect(self, ConfigDriver):
@@ -376,7 +376,7 @@ class DriverWrapperTests(DriverWrapperCommon):
         assert_equal(expected_driver, self.driver_wrapper.connect(maximize=False))
 
         # Check that the wrapper has been configured
-        assert_equal('firefox', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('firefox', self.driver_wrapper.config.get('Driver', 'type'))
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
         assert_equal('DEBUG', logging.getLevelName(logger.level))
 
@@ -384,7 +384,7 @@ class DriverWrapperTests(DriverWrapperCommon):
         # Mock data
         expected_driver = None
 
-        # Change browser to api and configure again
+        # Change driver type to api and configure again
         os.environ["Config_prop_filenames"] = os.path.join(self.root_path, 'conf', 'api-properties.cfg')
         self.driver_wrapper.configure()
 
@@ -393,23 +393,23 @@ class DriverWrapperTests(DriverWrapperCommon):
         assert_equal(expected_driver, self.driver_wrapper.connect(maximize=False))
 
         # Check that the wrapper has been configured
-        assert_equal('', self.driver_wrapper.config.get('Browser', 'browser'))
+        assert_equal('', self.driver_wrapper.config.get('Driver', 'type'))
         assert_equal('false', self.driver_wrapper.config.get('Jira', 'enabled'))
         logger = logging.getLogger('selenium.webdriver.remote.remote_connection')
         assert_equal('DEBUG', logging.getLevelName(logger.level))
 
     @data(*mobile_tests)
     @unpack
-    def test_is_mobile_test(self, browser, is_mobile, is_android, is_ios):
-        self.driver_wrapper.config.set('Browser', 'browser', browser)
+    def test_is_mobile_test(self, driver_type, is_mobile, is_android, is_ios):
+        self.driver_wrapper.config.set('Driver', 'type', driver_type)
         assert_equal(is_mobile, self.driver_wrapper.is_mobile_test())
         assert_equal(is_android, self.driver_wrapper.is_android_test())
         assert_equal(is_ios, self.driver_wrapper.is_ios_test())
 
     @data(*web_tests)
     @unpack
-    def test_is_web_test(self, browser, appium_app, appium_browser_name, is_web, is_android_web, is_ios_web):
-        self.driver_wrapper.config.set('Browser', 'browser', browser)
+    def test_is_web_test(self, driver_type, appium_app, appium_browser_name, is_web, is_android_web, is_ios_web):
+        self.driver_wrapper.config.set('Driver', 'type', driver_type)
         if appium_app is not None:
             self.driver_wrapper.config.set('AppiumCapabilities', 'app', appium_app)
         if appium_browser_name is not None:
@@ -418,8 +418,8 @@ class DriverWrapperTests(DriverWrapperCommon):
         assert_equal(is_android_web, self.driver_wrapper.is_android_web_test())
         assert_equal(is_ios_web, self.driver_wrapper.is_ios_web_test())
 
-    @data(*maximizable_browsers)
+    @data(*maximizable_drivers)
     @unpack
-    def test_is_maximizable(self, browser, is_maximizable):
-        self.driver_wrapper.config.set('Browser', 'browser', browser)
+    def test_is_maximizable(self, driver_type, is_maximizable):
+        self.driver_wrapper.config.set('Driver', 'type', driver_type)
         assert_equal(is_maximizable, self.driver_wrapper.is_maximizable())